+++ pyodide_build/common.py
@@ -7,13 +7,13 @@ TARGETPYTHON = ROOTDIR / '..' / 'cpython' / 'installs' / 'python-3.7.0'
 DEFAULTCFLAGS = ''
 DEFAULTLDFLAGS = ' '.join([
     '-O3',
-    '-s', "BINARYEN_METHOD='native-wasm'",
     '-Werror',
-    '-s', 'EMULATED_FUNCTION_POINTERS=1',
     '-s', 'EMULATE_FUNCTION_POINTER_CASTS=1',
     '-s', 'SIDE_MODULE=1',
     '-s', 'WASM=1',
-    '--memory-init-file', '0'
+    '--memory-init-file', '0',
+    '-s', 'LINKABLE=1',
+    '-s', 'EXPORT_ALL=1'
     ])